attendance Project/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.0
colorama==0.4.6
dlib==19.24.9
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.1
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
